refactor: modify navigaton file name #31
